+++ projects/ngx-enaio/src/lib/interfaces/enaio-group-info.ts
@@ -1,0 +1,23 @@
+
+/**
+ * enaio group information
+ */
+
+export interface EnaioGroupInfo {
+  /**
+   * Numeric User ID
+   */
+  id: number;
+  /**
+   * Group name
+   */
+  name: string;
+  /**
+   * Unique user identifier (32 characters long)
+   */
+  osguid: string;
+  /**
+   * group description
+   */
+  description: string;
+}