+++ projects/ngx-enaio/src/lib/interfaces/enaio-sendmail.ts
@@ -1,0 +1,23 @@
+
+/**
+ * Mail information
+ */
+
+export interface EnaioSendmail {
+  /**
+   * Sender
+   */
+  sender: string;
+  /**
+   * Receiver
+   */
+  receiver: string;
+  /**
+   * Mail subject
+   */
+  subject: string;
+  /**
+   * Mail body
+   */
+  text: string;
+}